Finalize the workflow to detect binodal decomposition.
Implemented a fitting procedure to obtain the densities of the two phases in the slab geometry.
The python code take in the densities and fit a double tanh, outputing the fit and the phase diagram.

Implemented simulation starting from a given configuration, so that can resume past simulations.
Simulations now leave a file containing the end configuration in their _video folders.

Correctly handled a premature exit by jumping to the end of the code and free up memories.

Allowed for two different velocity dependence on density in QSAPs.