Analyze and plot the density histograms using python.
Coarsen the histogram (important for QSAPs) and average over long time.
Fit a bimodal Gaussian to the normalized averaged histogram.
Create a phase diagram out of the fit results.
Print out the numerical values in the phase diagram into a text file.

Finalize the workflow with QSAPs as well.
Fixed minor bugs in the simulation codes.